refactor(api): split server.js into focused modules
Break the 999-line monolith into lib/config, auth, gcs, helpers, jobs,
and routes/{watermark,detect,download}. Also adds 4-minute process
timeout on C++ spawns, removes redundant rawResult from Firestore
detection writes, and documents the rate limiter IP-fallback trade-off.
